separated extractors classes in different files
