+++ src/deep_image_matching/extractors/__init__.py
@@ -1,0 +1,2 @@
+from .disk import DiskExtractor  # noqa: F401
+from .superpoint import SuperPointExtractor  # noqa: F401  # noqa: F401
